test(master-data): split 1082-line test_master_services.py by service (#113)
`services/master-data/tests/unit/test_master_services.py` had 13 test
classes (each service has a core class plus an *Additional class for
the second wave of coverage added in earlier work). Split into 7 files
— one per master service — combining each service's core+additional
classes:

- test_category_master_service.py        (116 lines)
- test_payment_master_service.py         (161 lines)
- test_staff_master_service.py           (159 lines)
- test_tax_master_service.py             (124 lines)
- test_settings_master_service.py        (261 lines)
- test_item_common_master_service.py     (163 lines)
- test_item_store_master_service.py      (237 lines)

Each new file imports only the documents/services it actually
references (test_item_store_master_service.py also pulls in
`ItemCommonMasterDocument` because TestItemStoreMasterService
constructs them in its mocks).

No test logic changed. Pass count unchanged: master-data unit tier
387/387 (was 387/387 before split).

Refs #113
